Add remaining fake method
